Add integration tests for class `ReviewAppClient`
This commits also changes the design of this class and add to it the responsibility to manage options and default values.

Thanks to the tests, many bugs or fragilities in either the implementation or the design were detected and fixed:
- the call to Scalingo API, GET /:app/operation/:id, was made at least one time too much
- internal method `pollOperationStatus` is now prefixed with '_' to show it is private method
- the same with `getClient` that is not intended to be used outside this class
- private methods were moved to the top (good practice to avoid JavaScript hoisting risks)
- the `executePoll` method inside `_pollOperationStatus` was reordered in a more functionnal & recursive style (with leaf cases first)
- the public `scale` does not resolve always anymore

This last point was a true design bug because some unexpected error could occurs over the HTTP response provided by the Scalingo API (ex: loss of Internet connection). The method now manage well expected and unexpected errors.

Finally, since `ReviewAppClient` is a wrapper around an external HTTP client, the more pertinent is to cover it by integration tests. It is why we here use Nock, a library that provide mechanisms to intercept HTTP requests and assert some scenarios or requests scopes. The setup of each test is a little bit fat, but they guaranty that the whole library is robust.